Move conda environment creation logic out of PythonWorkerFactory
